one-dimensional MVE (min. vol. ellipsoid)

generalized length of shortest-half sample

shorth(x, Alpha=0.5)

Arguments

x
data vector, no NAs

Alpha
minimum fraction of data to be covered by scale estimator. if Alpha == 0.5, the shorth is calculated

Value

a list, say L, with components

shorth

a 2-vector with endpoints of the shortest Alpha-sample

length.shorth

see previous return component L$shorth[2]-L$shorth[1]

midpt.shorth

mean(L[["shorth"]])

meanshorth

mean of values in the shorth, studied by Andrews et al (1972) as a location estimator

correction.parity.dep

correction factor to be applied to achieve approximate unbiasedness and diminish small-sample parity dependence; L["shorth"]] * L[["correction"]] is approximately unbiased for the Gaussian standard deviation, for 0 < Alpha < 1.

bias.correction.gau.5

correction factor to be applied along with correction.parity.dep when Alpha = .5; empirically derived bias correction useful for 10 < N < 2000 and possibly beyond. To use, divide: (L[["shorth"]] * L[["correction"]] / L[["bias.corr"]]) is approximately unbiased for Gaussian standard deviation, when Alpha=.5.

Alpha

coverage fraction used

References

Rousseeuw and Leroy, Stat Neer (1988), Gruebel, Ann Stat (1988)